requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.10
google-cloud-storage==2.18.2
qrcode[pil]==7.4.2
Pillow==10.4.0
celery==5.4.0
redis==5.0.8
python-dotenv==1.0.1
